Code/prompt_sentiment/backend/requirements.txt
flask_cors
flask==2.3.2
PyJWT==2.8.0
sqlalchemy==2.0.21
sqlalchemy_utils
flask_sqlalchemy>=2.1
flask_restx
flask-bcrypt>1
Werkzeug==2.3.3
psycopg2-binary
python-dotenv==1.0.0
openai==0.28.1
langchain==0.0.319
pandas
huggingface_hub
datasets
pytest
python-magic